Refactor clut4 conversion to its own file, use 8bpp resampling to better reuse code
