+++ source/psp/video_hardware.h
@@ -30,6 +30,7 @@ void GL_Upload16(int texture_index, const byte *data, int width, int height);
 int  GL_LoadTexture(const char *identifier, int width, int height, const byte *data, qboolean stretch_to_power_of_two, int filter, int mipmap_level);
 // CLUT4
 int GL_LoadTexture4(const char *identifier, unsigned int width, unsigned int height, const byte *data, int filter, qboolean swizzled);
+int GL_LoadTexture8to4(const char *identifier, unsigned int width, unsigned int height, const byte *data, const byte *pal, int filter);
 
 int GL_LoadTextureLM (const char *identifier, int width, int height, const byte *data, int bpp, int filter, qboolean update, int forcopy);
 int GL_LoadImages (const char *identifier, int width, int height, const byte *data, qboolean stretch_to_power_of_two, int filter, int mipmap_level, int bpp);
@@ -421,5 +422,6 @@ int 	      D_DrawParticleBuffered (psp_particle* vertices, particle2_t *pparticl
 
 extern int			zombie_skins[2][2];
 
-
 extern int faces_rejected, faces_checked, faces_clipped;
+
+void convert_8bpp_to_4bpp(const byte* indata, const byte* inpal, int width, int height, byte* outdata, byte* outpal);
